requirements.txt
Package                  Version            
------------------------ -------------------
absl-py                  0.9.0              
astor                    0.8.1              
astroid                  2.3.3              
attrs                    19.3.0             
backcall                 0.1.0              
bleach                   3.1.0              
cachetools               4.0.0              
certifi                  2019.11.28         
chardet                  3.0.4              
click                    7.1.1              
colorama                 0.4.3              
cycler                   0.10.0             
decorator                4.4.1              
defusedxml               0.6.0              
dlib                     19.7.0             
docutils                 0.16               
entrypoints              0.3                
face-recognition         1.3.0              
face-recognition-models  0.3.0              
gast                     0.2.2              
google-auth              1.10.2             
google-auth-oauthlib     0.4.1              
google-pasta             0.1.8              
graphviz                 0.13.2             
grpcio                   1.26.0             
h5py                     2.10.0             
idna                     2.8                
imageio                  2.6.1              
importlib-metadata       1.4.0              
ipykernel                5.1.3              
ipython                  7.11.1             
ipython-genutils         0.2.0              
ipywidgets               7.5.1              
isort                    4.3.21             
jedi                     0.15.2             
Jinja2                   2.10.3             
joblib                   0.14.1             
jsonschema               3.2.0              
jupyter                  1.0.0              
jupyter-client           5.3.4              
jupyter-console          6.1.0              
jupyter-core             4.6.1              
Keras-Applications       1.0.8              
Keras-Preprocessing      1.1.0              
Kivy                     1.11.1             
kivy-deps.glew           0.1.12             
kivy-deps.gstreamer      0.1.18             
kivy-deps.sdl2           0.1.23             
Kivy-examples            1.11.1             
Kivy-Garden              0.1.4              
kiwisolver               1.1.0              
lazy-object-proxy        1.4.3              
Markdown                 3.1.1              
MarkupSafe               1.1.1              
matplotlib               3.1.2              
mccabe                   0.6.1              
mistune                  0.8.4              
more-itertools           8.1.0              
nbconvert                5.6.1              
nbformat                 5.0.4              
networkx                 2.4                
notebook                 6.0.3              
numpy                    1.18.1             
oauthlib                 3.1.0              
opencv-python            4.1.2.30           
opencv-utils             0.0.2              
opt-einsum               3.1.0              
pandas                   0.25.3             
pandocfilters            1.4.2              
parso                    0.5.2              
pickleshare              0.7.5              
Pillow                   7.0.0              
pip                      20.0.1             
prometheus-client        0.7.1              
prompt-toolkit           3.0.2              
protobuf                 3.11.2             
pyasn1                   0.4.8              
pyasn1-modules           0.2.8              
pydot                    1.4.1              
pydot-ng                 2.0.0              
Pygments                 2.5.2              
pylint                   2.4.4              
pyparsing                2.4.6              
pypiwin32                223                
pyrsistent               0.15.7             
python-dateutil          2.8.1              
pytz                     2019.3             
PyWavelets               1.1.1              
pywin32                  227                
pywinpty                 0.5.7              
PyYAML                   5.3.1              
pyzmq                    18.1.1             
qtconsole                4.6.0              
requests                 2.22.0             
requests-oauthlib        1.3.0              
rsa                      4.0                
scikit-image             0.16.2             
scikit-learn             0.22.2.post1       
scipy                    1.4.1              
seaborn                  0.10.0             
Send2Trash               1.5.0              
setuptools               45.1.0.post20200119
six                      1.14.0             
tensorboard              1.15.0             
tensorflow-estimator     1.15.1             
tensorflow-gpu           1.15.0             
tensorflow-gpu-estimator 2.1.0              
termcolor                1.1.0              
terminado                0.8.3              
testpath                 0.4.4              
tornado                  6.0.3              
traitlets                4.3.3              
typed-ast                1.4.1              
urllib3                  1.25.8             
wcwidth                  0.1.8              
webencodings             0.5.1              
Werkzeug                 0.16.0             
wheel                    0.33.6             
widgetsnbextension       3.5.1              
wincertstore             0.2                
wrapt                    1.11.2             
zipp                     2.0.1              
